Make authorization for the BoothController

+++ app/Http/Controllers/Crew/BoothController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Crew;
 use App\Http\Controllers\Controller;
 use App\Models\Booth;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BoothController extends Controller
 {
@@ -13,7 +14,10 @@ class BoothController extends Controller
      */
     public function index()
     {
-        // Sort them so the companies that await approval appear on top
+        if (Auth::user()->cannot('viewAny', Booth::class)) {
+            abort(403);
+        }
+
         $booths = Booth::orderBy('is_approved')->paginate(15);
 
         return view('crew.booths.index', compact('booths'));
@@ -24,6 +28,10 @@ class BoothController extends Controller
      */
     public function create()
     {
+        if (Auth::user()->cannot('create', Booth::class)) {
+            abort(403);
+        }
+
         return view('crew.booths.create');
     }
 
@@ -32,6 +40,10 @@ class BoothController extends Controller
      */
     public function store(Request $request)
     {
+        if ($request->user()->cannot('create', Booth::class)) {
+            abort(403);
+        }
+
         $input = $request->validate([
             'company_id' => ['required', 'numeric'],
             'width' => ['required', 'numeric'],
@@ -52,6 +64,10 @@ class BoothController extends Controller
      */
     public function show(Booth $booth)
     {
+        if (Auth::user()->cannot('view', $booth)) {
+            abort(403);
+        }
+
         return view('crew.booths.show', compact('booth'));
     }
 
@@ -60,6 +76,10 @@ class BoothController extends Controller
      */
     public function approve(Request $request, Booth $booth)
     {
+        if (Auth::user()->cannot('approve', $booth)) {
+            abort(403);
+        }
+
         $validated = $request->validate([
             'approved' => 'required|boolean'
         ]);
@@ -78,6 +98,10 @@ class BoothController extends Controller
      */
     public function destroy(Booth $booth)
     {
+        if (Auth::user()->cannot('delete', $booth)) {
+            abort(403);
+        }
+
         $booth->delete();
 
         $template = 'You removed the booth of :company!';
